Separate isolated ordering tests
